+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 0.4.4 / ####-##-##
 ==================
 
+  * localStorage assertions
   * browser.setValue
 
 0.4.3 / 2018-03-19
 ==================
 
-  * localStore submodule
+  * localStorage submodule
   * browser.innerHtml and assertions
   * Query and QueryAll support optionally 2 parameters for subqueries
   * Assert text and not text now support arrays of expected strings